rename modules folder to methods
