+++ app/methods/Auth.js
